refactor: Rename DataGridView Extensions
This brings it in line with other extension methods
