+++ src/LessMsi.Gui/LessMsi.Gui.csproj
@@ -94,7 +94,7 @@
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
-    <Compile Include="Windows.Forms\DataGridViewExtensionMethods.cs" />
+    <Compile Include="Extensions\DataGridViewExtensions.cs" />
     <Compile Include="Windows.Forms\DisposableCursor.cs" />
     <Compile Include="Windows.Forms\ElevationButton.cs">
       <SubType>Component</SubType>
