+++ src/LessMsi.Gui/MainForm.cs
@@ -30,6 +30,7 @@ using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Windows.Forms;
+using LessMsi.Gui.Extensions;
 using LessMsi.Gui.Model;
 using LessMsi.Gui.Windows.Forms;
 using LessMsi.Msi;
